clip project! Keep only tokenizer and license.
