--- lib/pyxll/pyxll.cfg
@@ -1,10 +1,0 @@
-[PYXLL]
-pythonpath = ../../src
-modules = pycel.addin
-developer_mode = 1
-
-[LOG]
-verbosity = info
-format = %(asctime)s - %(levelname)s : %(message)s
-path = ./logs
-file = pyxll.%(date)s.log
