--- src/pycel/version.py
@@ -1,8 +1,0 @@
-# Store the version here so:
-# 1) we don't load dependencies by storing it in __init__.py
-# 2) we can import it in setup.py for the same reason
-# 3) we can import it into your module
-# See StackOverflow/458550 for more details
-
-# uses semantic versioning, http://semver.org
-__version__ = '0.0.1'
